+++ src/components/box/ProductItem.tsx
@@ -1,52 +1,12 @@
 import styled from '@emotion/styled';
-import { useState, useEffect } from 'react';
-import { useRecoilState } from 'recoil';
 import { CartIcon } from '../../assets';
-import type { CartItem, Product } from '../../types/types';
+import type { Product } from '../../types/types';
 import { Text } from '../common/Text/Text';
 import InputStepper from '../common/InputStepper/InputStepper';
-import { cartListState } from '../../store/atom';
+import useCartList from '../../hooks/useCartList';
 
 const ProductItem = ({ product }: { product: Product }) => {
-  const [cartList, setCartList] = useRecoilState(cartListState);
-
-  const existItemIndex = cartList.findIndex((cartItem) => cartItem.product.id === product.id);
-
-  const [quantity, setQuantity] = useState<number>(
-    existItemIndex !== -1 ? cartList[existItemIndex].quantity : 0,
-  );
-
-  const updateCartList = () => {
-    const newCartItem: CartItem = {
-      quantity,
-      product,
-    };
-
-    if (existItemIndex !== -1) {
-      const newCartList = cartList.slice();
-      newCartList.splice(existItemIndex, 1, newCartItem);
-      setCartList(newCartList);
-      return;
-    }
-
-    setCartList([...cartList, newCartItem]);
-  };
-
-  const deleteCartItem = () => {
-    if (existItemIndex !== -1) {
-      const newCartList = cartList.slice();
-      newCartList.splice(existItemIndex, 1);
-      setCartList(newCartList);
-    }
-  };
-
-  useEffect(() => {
-    if (quantity !== 0) {
-      updateCartList();
-      return;
-    }
-    deleteCartItem();
-  }, [quantity]);
+  const { quantity, setQuantity } = useCartList(product);
 
   return (
     <ProductWrapper>
